Refactor and separate the remote call related tracing test cases
