+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/observability/tracing/BaseTestCase.java
@@ -38,6 +38,7 @@ public class BaseTestCase extends BaseTest {
 
     protected static final String SERVER_CONNECTOR_NAME = "testobserve_listener";
     protected static final String MODULE_ID = "ballerina-test/testservices:0.0.1";
+    protected static final String COMMONS_FILE_NAME = "commons.bal";
 
     @BeforeGroups(value = "tracing-test", alwaysRun = true)
     private void setup() throws Exception {
