+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/observability/tracing/ResourceFunctionTestCase.java
@@ -22,6 +22,8 @@ import java.util.stream.Collectors;
 @Test(groups = "tracing-test")
 public class ResourceFunctionTestCase extends BaseTestCase {
     private static final String FILE_NAME = "02_resource_function.bal";
+    private static final String SERVICE_NAME = "testServiceOne";
+    private static final String BASE_URL = "http://localhost:9091";
 
     @DataProvider(name = "success-response-data-provider")
     public Object[][] getSuccessResponseData(){
@@ -34,13 +36,13 @@ public class ResourceFunctionTestCase extends BaseTestCase {
     @Test(dataProvider = "success-response-data-provider")
     public void testSimpleResourceCall(String resourceName, String requestPayload, String resourceFunctionPosition,
                                        String callerResponsePosition, String expectedResponsePayload) throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/" + resourceName,
+        HttpResponse httpResponse = HttpClientRequest.doPost(BASE_URL + "/" + SERVICE_NAME + "/" + resourceName,
                 requestPayload, Collections.emptyMap());
         Assert.assertEquals(httpResponse.getResponseCode(), 200);
         Assert.assertEquals(httpResponse.getData(), expectedResponsePayload);
         Thread.sleep(1000);
 
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", resourceName);
+        List<BMockSpan> spans = this.getFinishedSpans(SERVICE_NAME, resourceName);
         Assert.assertEquals(spans.stream()
                         .map(span -> span.getTags().get("src.position"))
                         .collect(Collectors.toSet()),
@@ -61,10 +63,10 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", resourceFunctionPosition),
                     new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/" + resourceName),
+                    new AbstractMap.SimpleEntry<>("http.url", "/" + SERVICE_NAME + "/" + resourceName),
                     new AbstractMap.SimpleEntry<>("http.method", "POST"),
                     new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
                     new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME)
             ));
@@ -83,7 +85,7 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", callerResponsePosition),
                     new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
                     new AbstractMap.SimpleEntry<>("connector_name", "ballerina/testobserve/Caller"),
                     new AbstractMap.SimpleEntry<>("action", "respond")
@@ -104,13 +106,13 @@ public class ResourceFunctionTestCase extends BaseTestCase {
     @Test(dataProvider = "error-response-data-provider")
     public void testResourceSingleSpanErrorResponse(String resourceName, String resourceFunctionPosition,
                                                     String expectedResponsePayload) throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/" + resourceName,
+        HttpResponse httpResponse = HttpClientRequest.doPost(BASE_URL + "/" + SERVICE_NAME + "/" + resourceName,
                 "15", Collections.emptyMap());
         Assert.assertEquals(httpResponse.getResponseCode(), 500);
         Assert.assertEquals(httpResponse.getData(), expectedResponsePayload);
         Thread.sleep(1000);
 
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", resourceName);
+        List<BMockSpan> spans = this.getFinishedSpans(SERVICE_NAME, resourceName);
         Assert.assertEquals(spans.stream()
                         .map(span -> span.getTags().get("src.position"))
                         .collect(Collectors.toSet()),
@@ -131,10 +133,10 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", resourceFunctionPosition),
                     new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/" + resourceName),
+                    new AbstractMap.SimpleEntry<>("http.url", "/" + SERVICE_NAME + "/" + resourceName),
                     new AbstractMap.SimpleEntry<>("http.method", "POST"),
                     new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
                     new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME),
                     new AbstractMap.SimpleEntry<>("error", "true")
@@ -147,7 +149,7 @@ public class ResourceFunctionTestCase extends BaseTestCase {
         return new Object[][] {
                 {"resourceSeven", FILE_NAME + ":87:5", FILE_NAME + ":89:30", FILE_NAME + ":96:20",
                         "Sum of numbers: 12"},
-                {"resourceThirteen", FILE_NAME + ":134:5", FILE_NAME + ":135:24", FILE_NAME + ":137:24",
+                {"resourceEight", FILE_NAME + ":100:5", FILE_NAME + ":101:24", FILE_NAME + ":103:24",
                         "Successfully executed"}
         };
     }
@@ -156,13 +158,13 @@ public class ResourceFunctionTestCase extends BaseTestCase {
     public void testResourceWithRemoteCall(String resourceName, String resourceFunctionPosition,
                                            String remoteCallPosition, String callerRespondPosition,
                                            String expectedPayload) throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/" + resourceName,
+        HttpResponse httpResponse = HttpClientRequest.doPost(BASE_URL + "/" + SERVICE_NAME + "/" + resourceName,
                 "", Collections.emptyMap());
         Assert.assertEquals(httpResponse.getResponseCode(), 200);
         Assert.assertEquals(httpResponse.getData(), expectedPayload);
         Thread.sleep(1000);
 
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", resourceName);
+        List<BMockSpan> spans = this.getFinishedSpans(SERVICE_NAME, resourceName);
         Assert.assertEquals(spans.stream()
                         .map(span -> span.getTags().get("src.position"))
                         .collect(Collectors.toSet()),
@@ -183,10 +185,10 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", resourceFunctionPosition),
                     new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/" + resourceName),
+                    new AbstractMap.SimpleEntry<>("http.url", "/" + SERVICE_NAME + "/" + resourceName),
                     new AbstractMap.SimpleEntry<>("http.method", "POST"),
                     new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
                     new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME)
             ));
@@ -205,7 +207,7 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", remoteCallPosition),
                     new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
                     new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
                     new AbstractMap.SimpleEntry<>("action", "callWithReturn")
@@ -225,351 +227,8 @@ public class ResourceFunctionTestCase extends BaseTestCase {
                     new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
                     new AbstractMap.SimpleEntry<>("src.position", callerRespondPosition),
                     new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
+                    new AbstractMap.SimpleEntry<>("service", SERVICE_NAME),
                     new AbstractMap.SimpleEntry<>("resource", resourceName),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina/testobserve/Caller"),
-                    new AbstractMap.SimpleEntry<>("action", "respond")
-            ));
-        });
-    }
-
-    @Test
-    public void testResourceWithNestedRemoteCalls() throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/resourceEight",
-                "", Collections.emptyMap());
-        Assert.assertEquals(httpResponse.getResponseCode(), 200);
-        Assert.assertEquals(httpResponse.getData(), "Invocation Successful");
-        Thread.sleep(1000);
-
-        final String span1Position = FILE_NAME + ":100:5";
-        final String span2Position = FILE_NAME + ":101:9";
-        final String span3Position = "commons.bal:36:9";
-        final String span4Position = FILE_NAME + ":102:20";
-
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", "resourceEight");
-        Assert.assertEquals(spans.stream()
-                        .map(span -> span.getTags().get("src.position"))
-                        .collect(Collectors.toSet()),
-                new HashSet<>(Arrays.asList(span1Position, span2Position, span3Position, span4Position)));
-        Assert.assertEquals(spans.stream().filter(bMockSpan -> bMockSpan.getParentId() == 0).count(), 1);
-
-        Optional<BMockSpan> span1 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span1Position))
-                .findFirst();
-        Assert.assertTrue(span1.isPresent());
-        long traceId = span1.get().getTraceId();
-        span1.ifPresent(span -> {
-            Assert.assertTrue(spans.stream().noneMatch(mockSpan -> mockSpan.getTraceId() == traceId
-                    && mockSpan.getSpanId() == span.getParentId()));
-            Assert.assertEquals(span.getOperationName(), "resourceEight");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "server"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span1Position),
-                    new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/resourceEight"),
-                    new AbstractMap.SimpleEntry<>("http.method", "POST"),
-                    new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEight"),
-                    new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME)
-            ));
-        });
-
-        Optional<BMockSpan> span2 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span2Position))
-                .findFirst();
-        Assert.assertTrue(span2.isPresent());
-        span2.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(),
-                    "ballerina-test/testservices/MockClient:callAnotherRemoteFunction");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span2Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEight"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
-                    new AbstractMap.SimpleEntry<>("action", "callAnotherRemoteFunction")
-            ));
-        });
-
-        Optional<BMockSpan> span3 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span3Position))
-                .findFirst();
-        Assert.assertTrue(span3.isPresent());
-        span3.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span2.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina-test/testservices/MockClient:callWithNoReturn");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span3Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEight"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
-                    new AbstractMap.SimpleEntry<>("action", "callWithNoReturn")
-            ));
-        });
-
-        Optional<BMockSpan> span4 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span4Position))
-                .findFirst();
-        Assert.assertTrue(span4.isPresent());
-        span4.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina/testobserve/Caller:respond");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span4Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEight"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina/testobserve/Caller"),
-                    new AbstractMap.SimpleEntry<>("action", "respond")
-            ));
-        });
-    }
-
-    @DataProvider(name = "remote-error-return-data-provider")
-    public Object[][] getRemoteErrorReturnData(){
-        return new Object[][] {
-                {"resourceNine", FILE_NAME + ":106:5", FILE_NAME + ":107:15"},
-                {"resourceTen", FILE_NAME + ":112:5", FILE_NAME + ":113:20"}
-        };
-    }
-
-    @Test(dataProvider = "remote-error-return-data-provider")
-    public void testRemoteFunctionErrorReturn(String resourceName, String resourceFunctionPosition,
-                                              String remoteCallPosition) throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/" + resourceName,
-                "", Collections.emptyMap());
-        Assert.assertEquals(httpResponse.getResponseCode(), 500);
-        Assert.assertEquals(httpResponse.getData(), "Test Error");
-        Thread.sleep(1000);
-
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", resourceName);
-        Assert.assertEquals(spans.stream()
-                        .map(span -> span.getTags().get("src.position"))
-                        .collect(Collectors.toSet()),
-                new HashSet<>(Arrays.asList(resourceFunctionPosition, remoteCallPosition)));
-        Assert.assertEquals(spans.stream().filter(bMockSpan -> bMockSpan.getParentId() == 0).count(), 1);
-
-        Optional<BMockSpan> span1 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), resourceFunctionPosition))
-                .findFirst();
-        Assert.assertTrue(span1.isPresent());
-        long traceId = span1.get().getTraceId();
-        span1.ifPresent(span -> {
-            Assert.assertTrue(spans.stream().noneMatch(mockSpan -> mockSpan.getTraceId() == traceId
-                    && mockSpan.getSpanId() == span.getParentId()));
-            Assert.assertEquals(span.getOperationName(), resourceName);
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "server"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", resourceFunctionPosition),
-                    new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/" + resourceName),
-                    new AbstractMap.SimpleEntry<>("http.method", "POST"),
-                    new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", resourceName),
-                    new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME),
-                    new AbstractMap.SimpleEntry<>("error", "true")
-            ));
-        });
-
-        Optional<BMockSpan> span2 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), remoteCallPosition))
-                .findFirst();
-        Assert.assertTrue(span2.isPresent());
-        span2.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina-test/testservices/MockClient:callWithErrorReturn");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", remoteCallPosition),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", resourceName),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
-                    new AbstractMap.SimpleEntry<>("action", "callWithErrorReturn"),
-                    new AbstractMap.SimpleEntry<>("error", "true")
-            ));
-        });
-    }
-
-    @Test
-    public void testResourceWithIgnoredErrorReturnInRemoteCall() throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/resourceEleven",
-                "", Collections.emptyMap());
-        Assert.assertEquals(httpResponse.getResponseCode(), 200);
-        Assert.assertEquals(httpResponse.getData(), "Invocation Successful");
-        Thread.sleep(1000);
-
-        final String span1Position = FILE_NAME + ":118:5";
-        final String span2Position = FILE_NAME + ":119:19";
-        final String span3Position = FILE_NAME + ":120:20";
-
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", "resourceEleven");
-        Assert.assertEquals(spans.stream()
-                        .map(span -> span.getTags().get("src.position"))
-                        .collect(Collectors.toSet()),
-                new HashSet<>(Arrays.asList(span1Position, span2Position, span3Position)));
-        Assert.assertEquals(spans.stream().filter(bMockSpan -> bMockSpan.getParentId() == 0).count(), 1);
-
-        Optional<BMockSpan> span1 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span1Position))
-                .findFirst();
-        Assert.assertTrue(span1.isPresent());
-        long traceId = span1.get().getTraceId();
-        span1.ifPresent(span -> {
-            Assert.assertTrue(spans.stream().noneMatch(mockSpan -> mockSpan.getTraceId() == traceId
-                    && mockSpan.getSpanId() == span.getParentId()));
-            Assert.assertEquals(span.getOperationName(), "resourceEleven");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "server"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span1Position),
-                    new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/resourceEleven"),
-                    new AbstractMap.SimpleEntry<>("http.method", "POST"),
-                    new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEleven"),
-                    new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME)
-            ));
-        });
-
-        Optional<BMockSpan> span2 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span2Position))
-                .findFirst();
-        Assert.assertTrue(span2.isPresent());
-        span2.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina-test/testservices/MockClient:callWithErrorReturn");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span2Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEleven"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
-                    new AbstractMap.SimpleEntry<>("action", "callWithErrorReturn"),
-                    new AbstractMap.SimpleEntry<>("error", "true")
-            ));
-        });
-
-        Optional<BMockSpan> span3 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span3Position))
-                .findFirst();
-        Assert.assertTrue(span3.isPresent());
-        span3.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina/testobserve/Caller:respond");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span3Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceEleven"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina/testobserve/Caller"),
-                    new AbstractMap.SimpleEntry<>("action", "respond")
-            ));
-        });
-    }
-
-    @Test
-    public void testResourceWithTrappedPanic() throws Exception {
-        HttpResponse httpResponse = HttpClientRequest.doPost("http://localhost:9091/testServiceOne/resourceTwelve",
-                "", Collections.emptyMap());
-        Assert.assertEquals(httpResponse.getResponseCode(), 200);
-        Assert.assertEquals(httpResponse.getData(), "Successfully trapped panic: Test Error");
-        Thread.sleep(1000);
-
-        final String span1Position = FILE_NAME + ":124:5";
-        final String span2Position = FILE_NAME + ":125:24";
-        final String span3Position = FILE_NAME + ":127:24";
-
-        List<BMockSpan> spans = this.getFinishedSpans("testServiceOne", "resourceTwelve");
-        Assert.assertEquals(spans.stream()
-                        .map(span -> span.getTags().get("src.position"))
-                        .collect(Collectors.toSet()),
-                new HashSet<>(Arrays.asList(span1Position, span2Position, span3Position)));
-        Assert.assertEquals(spans.stream().filter(bMockSpan -> bMockSpan.getParentId() == 0).count(), 1);
-
-        Optional<BMockSpan> span1 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span1Position))
-                .findFirst();
-        Assert.assertTrue(span1.isPresent());
-        long traceId = span1.get().getTraceId();
-        span1.ifPresent(span -> {
-            Assert.assertTrue(spans.stream().noneMatch(mockSpan -> mockSpan.getTraceId() == traceId
-                    && mockSpan.getSpanId() == span.getParentId()));
-            Assert.assertEquals(span.getOperationName(), "resourceTwelve");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "server"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span1Position),
-                    new AbstractMap.SimpleEntry<>("src.entry_point.resource", "true"),
-                    new AbstractMap.SimpleEntry<>("http.url", "/testServiceOne/resourceTwelve"),
-                    new AbstractMap.SimpleEntry<>("http.method", "POST"),
-                    new AbstractMap.SimpleEntry<>("protocol", "http"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceTwelve"),
-                    new AbstractMap.SimpleEntry<>("connector_name", SERVER_CONNECTOR_NAME)
-            ));
-        });
-
-        Optional<BMockSpan> span2 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span2Position))
-                .findFirst();
-        Assert.assertTrue(span2.isPresent());
-        span2.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina-test/testservices/MockClient:callWithPanic");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span2Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceTwelve"),
-                    new AbstractMap.SimpleEntry<>("connector_name", "ballerina-test/testservices/MockClient"),
-                    new AbstractMap.SimpleEntry<>("action", "callWithPanic"),
-                    new AbstractMap.SimpleEntry<>("error", "true")
-            ));
-        });
-
-        Optional<BMockSpan> span3 = spans.stream()
-                .filter(bMockSpan -> Objects.equals(bMockSpan.getTags().get("src.position"), span3Position))
-                .findFirst();
-        Assert.assertTrue(span3.isPresent());
-        span3.ifPresent(span -> {
-            Assert.assertEquals(span.getTraceId(), traceId);
-            Assert.assertEquals(span.getParentId(), span1.get().getSpanId());
-            Assert.assertEquals(span.getOperationName(), "ballerina/testobserve/Caller:respond");
-            Assert.assertEquals(span.getTags(), toMap(
-                    new AbstractMap.SimpleEntry<>("span.kind", "client"),
-                    new AbstractMap.SimpleEntry<>("src.module", MODULE_ID),
-                    new AbstractMap.SimpleEntry<>("src.position", span3Position),
-                    new AbstractMap.SimpleEntry<>("src.remote", "true"),
-                    new AbstractMap.SimpleEntry<>("service", "testServiceOne"),
-                    new AbstractMap.SimpleEntry<>("resource", "resourceTwelve"),
                     new AbstractMap.SimpleEntry<>("connector_name", "ballerina/testobserve/Caller"),
                     new AbstractMap.SimpleEntry<>("action", "respond")
             ));
