+++ tests/jballerina-integration-test/src/test/resources/testng.xml
@@ -209,6 +209,7 @@
             <class name="org.ballerinalang.test.observability.tracing.BaseTestCase"/>
             <class name="org.ballerinalang.test.observability.tracing.MainFunctionTestCase"/>
             <class name="org.ballerinalang.test.observability.tracing.ResourceFunctionTestCase"/>
+            <class name="org.ballerinalang.test.observability.tracing.RemoteCallTestCase"/>
 <!--            <class name="org.ballerinalang.test.observability.tracing.TracingBaseTest"/>-->
 <!--            <class name="org.ballerinalang.test.observability.tracing.OOTBTracingTestCase"/>-->
 <!--            <class name="org.ballerinalang.test.observability.tracing.CustomTracingTestCase"/>-->
